LayoutTests/http/tests/site-isolation/inspector/debugger/setBreakpoint-cross-origin-iframe.html

<script src="../../../inspector/resources/inspector-test.js"></script>
<script src="resources/site-isolation-debugger-test-utilities.js"></script>
<script>
let iframe;

function addCrossOriginIFrame() {
    let crossOriginHost = location.hostname === "localhost" ? "127.0.0.1" : "localhost";
    iframe = document.createElement("iframe");
    iframe.src = `http://${crossOriginHost}:8000/site-isolation/inspector/debugger/resources/debugger-frame.html`;
    document.body.appendChild(iframe);
}

function test() {
    let suite = InspectorTest.createAsyncSuite("SiteIsolation.Debugger.SetBreakpoint");

    let crossOriginTarget;
    let breakpointId;
    let crossOriginHost;

    suite.addTestCase({
        name: "SiteIsolation.Debugger.SetBreakpoint.Setup",
        description: "Set up cross-origin iframe.",
        async test() {
            let whenTargetAddedPromise = WI.targetManager.awaitEvent(WI.TargetManager.Event.TargetAdded);
            InspectorTest.evaluateInPage("addCrossOriginIFrame()");
            let event = await whenTargetAddedPromise;
            crossOriginTarget = event.data.target;

            InspectorTest.assert(crossOriginTarget instanceof WI.FrameTarget);

            // Determine the cross-origin host for breakpoint URL.
            let hostnameResponse = await crossOriginTarget.RuntimeAgent.evaluate.invoke({expression: "location.hostname", objectGroup: "test", returnByValue: true});
            crossOriginHost = hostnameResponse.result.value;

            // Wait for the frame's scripts to load.
            await waitForExpressionInTarget(crossOriginTarget, "typeof breakpointTarget === 'function'");
        }
    });

    suite.addTestCase({
        name: "SiteIsolation.Debugger.SetBreakpoint.SetByUrl",
        description: "Setting a breakpoint by URL should resolve in the frame target's script.",
        async test() {
            // Line 8 (0-indexed) in debugger-frame-script.js is: let localVar = x * 2;
            let scriptUrl = `http://${crossOriginHost}:8000/site-isolation/inspector/debugger/resources/debugger-frame-script.js`;
            let result = await crossOriginTarget.DebuggerAgent.setBreakpointByUrl.invoke({
                lineNumber: 8,
                url: scriptUrl,
            });

            InspectorTest.expectNotNull(result.breakpointId, "Breakpoint should be set successfully.");
            InspectorTest.expectGreaterThan(result.locations.length, 0, "Breakpoint should have at least one resolved location.");

            breakpointId = result.breakpointId;
        }
    });

    suite.addTestCase({
        name: "SiteIsolation.Debugger.SetBreakpoint.HitsBreakpoint",
        description: "Calling the function should hit the breakpoint.",
        async test() {
            let whenPausedPromise = WI.debuggerManager.awaitEvent(WI.DebuggerManager.Event.Paused);

            // Call the function that has the breakpoint.
            crossOriginTarget.RuntimeAgent.evaluate.invoke({expression: "breakpointTarget(5)", objectGroup: "test"});

            await whenPausedPromise;

            InspectorTest.expectThat(WI.debuggerManager.paused, "Debugger should be paused at breakpoint.");

            let targetData = WI.debuggerManager.dataForTarget(crossOriginTarget);
            let topCallFrame = targetData.stackTrace.callFrames[0];
            InspectorTest.expectEqual(topCallFrame.functionName, "breakpointTarget", "Should be paused in 'breakpointTarget'.");

            // Evaluate local variable.
            let result = await crossOriginTarget.DebuggerAgent.evaluateOnCallFrame.invoke({
                callFrameId: topCallFrame.id,
                expression: "x",
                objectGroup: "test",
                returnByValue: true,
            });
            InspectorTest.expectEqual(result.result.value, 5, "Parameter 'x' should have value 5.");

            // Resume.
            await WI.debuggerManager.resume();
        }
    });

    suite.addTestCase({
        name: "SiteIsolation.Debugger.SetBreakpoint.RemoveBreakpoint",
        description: "Removing the breakpoint should allow execution without pausing.",
        async test() {
            await crossOriginTarget.DebuggerAgent.removeBreakpoint(breakpointId);

            let didPause = false;
            let listener = WI.debuggerManager.addEventListener(WI.DebuggerManager.Event.Paused, () => { didPause = true; });

            let result = await crossOriginTarget.RuntimeAgent.evaluate.invoke({expression: "breakpointTarget(10)", objectGroup: "test", returnByValue: true});
            InspectorTest.expectEqual(result.result.value, 21, "Function should return correct result (10 * 2 + 1 = 21).");

            // Give a brief moment to ensure no pause event fires.
            await new Promise((resolve) => setTimeout(resolve, 100));

            InspectorTest.expectFalse(didPause, "Debugger should not pause after breakpoint removal.");

            WI.debuggerManager.removeEventListener(WI.DebuggerManager.Event.Paused, listener);
        }
    });

    suite.runTestCasesAndFinish();
}
</script>

<body onload="runTest()">
<p>Test that breakpoints can be set and hit in cross-origin iframe scripts.</p>
</body>
